pipelines/requirements.txt
aadict==0.2.3
amundsen-databuilder==3.3.2
asn1crypto==1.4.0
asset==0.6.13
azure-common==1.1.25
azure-core==1.8.2
azure-storage-blob==12.5.0
boto3==1.16.6
botocore==1.19.6
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
colored==1.1.0
cryptography==3.2
elasticsearch==6.8.1
future==0.18.2
globre==0.1.5
google-api-core==1.23.0
google-api-python-client==1.10.0
google-auth==1.22.1
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
google-cloud-bigquery==2.2.0
google-cloud-core==1.4.3
google-cloud-datacatalog==2.0.0
google-cloud-spanner==1.19.0
google-crc32c==1.0.0
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
grpcio==1.32.0
grpcio-gcp==0.2.2
httplib2==0.18.1
idna==2.9
isodate==0.6.0
Jinja2==2.11.2
jmespath==0.10.0
msrest==0.6.19
neo4j-driver==1.7.6
neobolt==1.7.17
neotime==1.7.4
numpy==1.19.2
oauthlib==3.1.0
oscrypto==1.2.1
pandas==1.1.1
proto-plus==1.11.0
protobuf==3.13.0
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybigquery==0.4.15
pycparser==2.20
pycryptodomex==3.9.8
PyHive==0.6.2
pyhocon==0.3.56
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.6
s3transfer==0.3.3
six==1.15.0
snowflake-connector-python==2.3.2
snowflake-sqlalchemy==1.2.3
SQLAlchemy==1.3.19
statsd==3.3.0
termcolor==1.1.0
Unidecode==1.1.1
uritemplate==3.0.1
urllib3==1.25.11
